delete original message on post, cleanup code
