+++ src/config.py
@@ -1,0 +1,4 @@
+# config.py
+
+# Discord Bot Token
+BOT_TOKEN = 'INSERT BOT TOKEN HERE'
